test: build a macro shim to authenticate channel joins
